+++ src/include/kdbglobal.h
@@ -1,0 +1,63 @@
+/**
+ * @file
+ *
+ * @brief Defines for global plugins
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ */
+
+#ifndef ELEKTRA_KDBGLOBAL_H
+#define ELEKTRA_KDBGLOBAL_H
+
+#include <kdb.h>
+#include <kdbplugin.h>
+
+/**
+ * Helper for identifying global plugin positions
+ */
+
+#define FOREACH_POSITION(POSITION) \
+	POSITION(PREROLLBACK) \
+	POSITION(ROLLBACK) \
+	POSITION(POSTROLLBACK) \
+	POSITION(GETRESOLVER) \
+	POSITION(PREGETSTORAGE) \
+	POSITION(GETSTORAGE) \
+	POSITION(POSTGETSTORAGE) \
+	POSITION(SETRESOLVER) \
+	POSITION(POSTGETCLEANUP) \
+	POSITION(PRESETSTORAGE) \
+	POSITION(SETSTORAGE) \
+	POSITION(PRESETCLEANUP) \
+	POSITION(PRECOMMIT) \
+	POSITION(COMMIT) \
+	POSITION(POSTCOMMIT) \
+	POSITION(NR_GLOBAL_POSITIONS)
+
+#define FOREACH_SUBPOSITION(SUBPOSITION) \
+	SUBPOSITION(INIT) \
+	SUBPOSITION(MAXONCE) \
+	SUBPOSITION(FOREACH) \
+	SUBPOSITION(DEINIT) \
+	SUBPOSITION(NR_GLOBAL_SUBPOSITIONS)
+
+#define GENERATE_ENUM(ENUM) ENUM,
+#define GENERATE_STRING(STRING) #STRING,
+
+typedef enum {
+	FOREACH_POSITION(GENERATE_ENUM)
+} GlobalpluginPositions;
+
+typedef enum {
+	FOREACH_SUBPOSITION(GENERATE_ENUM)
+} GlobalpluginSubPositions;
+
+static const char *GlobalpluginPositionsStr[] = {
+		FOREACH_POSITION(GENERATE_STRING)
+};
+
+static const char *GlobalpluginSubPositionsStr[] = {
+		FOREACH_SUBPOSITION(GENERATE_STRING)
+};
+
+#endif //ELEKTRA_KDBGLOBAL_H
