+++ src/libs/elektra/kdb.c
@@ -145,37 +145,6 @@ void elektraRemoveMetaData (Key * key, const char * searchfor)
 }
 
 /**
- * @internal
- * Helper functions to execute global plugins
- */
-void elektraGlobalGet (KDB * handle, KeySet * ks, Key * parentKey, int position, int subPosition)
-{
-	Plugin * plugin;
-	if (handle && (plugin = handle->globalPlugins[position][subPosition]))
-	{
-		plugin->kdbGet (plugin, ks, parentKey);
-	}
-}
-
-void elektraGlobalSet (KDB * handle, KeySet * ks, Key * parentKey, int position, int subPosition)
-{
-	Plugin * plugin;
-	if (handle && (plugin = handle->globalPlugins[position][subPosition]))
-	{
-		plugin->kdbSet (plugin, ks, parentKey);
-	}
-}
-
-void elektraGlobalError (KDB * handle, KeySet * ks, Key * parentKey, int position, int subPosition)
-{
-	Plugin * plugin;
-	if (handle && (plugin = handle->globalPlugins[position][subPosition]))
-	{
-		plugin->kdbError (plugin, ks, parentKey);
-	}
-}
-
-/**
  * @brief Bootstrap, first phase with fallback
  * @internal
  *
